Cleanup ARM connector and extend logging
